fix(m6): address multi-agent code-review findings
P1s:
- KV increment no longer truncates floats on SQLite (CAST AS REAL, not
  INTEGER) — dual-dialect parity for float counters (+ regression test).
- Password-gate the runtime API: a shared password-protected canvas's
  KV/files/me stay closed without a valid gate grant (§12.0 #3) (+ test).
- Wire usage_events pruneBefore into the purge sweep (KTD-7 retention;
  was never called → unbounded growth).
- SDK files.upload() returns an absolute content url (server's was
  root-relative → broke subdomain mode) (+ test).

P2s:
- Path-mode Referer check is now segment-boundary, not substring
  (app vs app-evil) (+ test); standardized runtime-API error envelopes
  to {code} (SCREAMING_SNAKE); SDK subdomain apiBase preserves the port
  (+ test); serve-sdk no longer sticky-caches a failed load; countFiles
  aggregate replaces full files.list() in the usage endpoint; extracted
  shared requireCanvas helper (dedup); FileMeta.createdAt added.
- Added route-level regression tests: NOT_NUMERIC, KEY_LIMIT, files
  backend-off, files quota 409.

+++ apps/server/src/db/repositories/kv.ts
@@ -113,9 +113,11 @@ export function kvRepository(client: DbClient) {
       if (existing !== null && typeof existing !== "number") throw new KvNotNumericError();
 
       const now = Date.now();
-      // Dialect-specific atomic numeric expression on the row's own value.
+      // Dialect-specific atomic numeric expression on the row's own value. Use a
+      // REAL cast on SQLite (NOT INTEGER) so fractional values aren't truncated —
+      // keeps both dialects in lockstep for float counters (pg uses ::numeric).
       const nextExpr = isSqlite
-        ? sql`CAST(${t.value} AS INTEGER) + ${by}`
+        ? sql`CAST(${t.value} AS REAL) + ${by}`
         : sql`to_jsonb((${t.value}::text::numeric) + ${by})`;
       const rows = (await db
         .insert(t)
